refactor: keep Prometheus specifics out of the data plugin
Move the PromQL step helpers back to query_enhancements/common and expose the
module through extraPublicDirs so explore's bundle can import it, and read the
step resolution off the data frame meta like profile already does, dropping the
convertResult passthrough. Reuse the query_enhancements PromQL option and step
resolution types in explore instead of redeclaring them.
