+++ src/plugins/data/common/data_frames/utils.ts
@@ -149,10 +149,6 @@ export const convertResult = ({
     (searchResponse as any).truncation = data.meta.truncation;
   }
 
-  if (data.meta?.stepResolution) {
-    (searchResponse as any).stepResolution = data.meta.stepResolution;
-  }
-
   if (data.hasOwnProperty('aggs')) {
     const dataWithAggs = data as IDataFrameWithAggs;
     if (!dataWithAggs.aggs) {
