+++ src/plugins/explore/public/application/pages/metrics/query_panel/metrics_query_options.tsx
@@ -17,7 +17,10 @@ import {
   EuiText,
 } from '@elastic/eui';
 import { PerQueryOptions } from '../../../utils/languages';
-import { formatPromDuration, parseStepIntervalSeconds } from '../../../../../../data/common';
+import {
+  formatPromDuration,
+  parseStepIntervalSeconds,
+} from '../../../../../../query_enhancements/common/prom_step';
 
 export interface RowStepReadout {
   stepLabel: string;
