+++ src/plugins/explore/public/application/utils/state_management/actions/query_actions.ts
@@ -704,6 +704,7 @@ const executeQueryBase = async (
       elapsedMs: inspectorRequest.getTime()!,
       fieldSchema: searchSource.getDataFrame()?.schema,
       profile: searchSource.getDataFrame()?.meta?.profile,
+      stepResolution: searchSource.getDataFrame()?.meta?.stepResolution,
     };
 
     if (isHistogramQuery && effectiveHistogramConfig) {
