+++ src/plugins/query_enhancements/common/index.ts
@@ -3,6 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+// Do not re-export ./prom_step here. Explore's browser bundle imports it by path
+// because that path is the only entry listed in extraPublicDirs; routing it through
+// this barrel makes the optimizer reject it as a non-public export.
 export * from './constants';
 export * from './prom_legend';
 export * from './types';
