+++ src/plugins/query_enhancements/opensearch_dashboards.json
@@ -5,6 +5,7 @@
   "ui": true,
   "requiredPlugins": ["data", "opensearchDashboardsReact", "opensearchDashboardsUtils", "savedObjects", "uiActions"],
   "optionalPlugins": ["dataSource", "usageCollection"],
+  "extraPublicDirs": ["common/prom_step"],
   "configPath": ["queryEnhancements"]
 }
